solutions/reverse_integer.py
def reverse(self, x: int) -> int:
    """
    Reverses the digits of a signed 32-bit integer.

    Parameters:
    x (int): The integer to reverse.

    Returns:
    int: The reversed integer if it falls within the signed 32-bit range [-2^31, 2^31 - 1].
         Returns 0 if the reversed integer overflows.

    Examples:
    >>> reverse(None, 123)
    321
    >>> reverse(None, -123)
    -321
    >>> reverse(None, 120)
    21
    >>> reverse(None, 0)
    0
    """
    # Define the range for signed 32-bit integers
    INT_MIN, INT_MAX = -2**31, 2**31 - 1

    # Determine if the number is negative
    negative = x < 0

    # Work with the absolute value of x
    x = abs(x)

    # Reverse the digits using string manipulation
    reversed_x = int(str(x)[::-1])

    # Restore the sign if the number was negative
    if negative:
        reversed_x = -reversed_x

    # Check for overflow
    if reversed_x < INT_MIN or reversed_x > INT_MAX:
        return 0

    return reversed_x
